+++ plugins/dashboards/src/utils/dashboard.ts
@@ -1,7 +1,7 @@
 import { gridSpans } from '@patternfly/react-core';
 
-import { IDashboard, IPlaceholders, IPluginDefaults, IPluginTimes, IReference } from '@kobsio/plugin-core';
-import { IDashboardsOptions, IVariableValues } from './interfaces';
+import { IDashboard, IDashboardReference, IPlaceholders, IPluginDefaults } from '@kobsio/plugin-core';
+import { IDashboardsOptions } from './interfaces';
 
 // toGridSpans is used to convert the provided col and row span value to the corresponding gridSpans value, so that it
 // can be used within the Patternfly Grid component. The function requires a default value which is 12 for columns and
@@ -31,61 +31,14 @@ export const rowHeight = (rowSize: number | undefined, rowSpan: number | undefin
   return `${rowSize * 150}px`;
 };
 
-// IVariables is a map of variable names with the current value. This interface should only be used by the interpolate
-// function, to convert a given array of variables to the format, which is required by the function.
-interface IVariables {
-  [key: string]: string;
-}
-
-// interpolate is used to replace the variables in a given string with the current value for this variable. Before we
-// can replace the variables in a string we have to convert the array of variables to a map of variable names and there
-// value.
-// The default interpolator/delimiter is "{%" and "%}", so that it doesn't conflict with the delimiter used for the
-// placeholder. We can not use the same, because the are replaced at different points in our app logic.
-// See: https://stackoverflow.com/a/57598892/4104109
-export const interpolate = (
-  str: string,
-  variables: IVariableValues[],
-  times: IPluginTimes,
-  interpolator: string[] = ['{%', '%}'],
-): string => {
-  const vars: IVariables = {};
-
-  for (const variable of variables) {
-    vars[variable.name] = variable.value;
-  }
-
-  vars['__timeStart'] = `${times.timeStart}`;
-  vars['__timeEnd'] = `${times.timeEnd}`;
-
-  return str
-    .split(interpolator[0])
-    .map((s1, i) => {
-      if (i === 0) {
-        return s1;
-      }
-
-      const s2 = s1.split(interpolator[1]);
-      if (s1 === s2[0]) {
-        return interpolator[0] + s2[0];
-      }
-
-      if (s2.length > 1) {
-        s2[0] =
-          s2[0] && vars.hasOwnProperty(s2[0].trim().substring(1))
-            ? vars[s2[0].trim().substring(1)]
-            : interpolator.join(` ${s2[0]} `);
-      }
-
-      return s2.join('');
-    })
-    .join('');
-};
-
 // getOptionsFromSearch is used to parse the given search location and return is as options for Prometheus. This is
 // needed, so that a user can explore his Prometheus data from a chart. When the user selects the explore action, we
 // pass him to this page and pass the data via the URL parameters.
-export const getInitialOptions = (search: string, references: IReference[], useDrawer: boolean): IDashboardsOptions => {
+export const getInitialOptions = (
+  search: string,
+  references: IDashboardReference[],
+  useDrawer: boolean,
+): IDashboardsOptions => {
   const params = new URLSearchParams(search);
   const dashboard = params.get('dashboard');